+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added transparent support for bearer tokens [#341](https://github.com/vmware/go-vcloud-director/pull/341)
 * Added transparent connection using `cloudapi/1.0.0/sessions` when access through `api/sessions` is disabled
 * Added functions `edge.GetLbAppRules`, `edge.GetLbServerPools`, `edge.GetLbAppProfiles`, `edge.GetNsxvNatRules`, `client.GetOrgList`
+* Exported private function `client.maxSupportedVersion` to `client.MaxSupportedVersion`
 
 BREAKING CHANGES: